ticket-summarizer/requirements.txt
Flask~=3.1.2

google-genai~=1.39.1